+++ src/ByteDev.Azure.KeyVault/Secrets/IKeyVaultSecretClient.cs
@@ -23,22 +23,34 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be true if the secret exists; otherwise false.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task<bool> ExistsAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Retrieves all secrets.
         /// </summary>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be a list of secrets.</returns>
         Task<IList<KeyVaultSecret>> GetAllAsync(CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Retrieves all secrets for a particular section. For example the section name for secret names:
+        /// "MySection--Secret1" and "MySection--Secret2" is "MySection".
+        /// </summary>
+        /// <param name="sectionName">Section name.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The task object representing the asynchronous operation. Result will be a list of secrets.</returns>
+        Task<IList<KeyVaultSecret>> GetSectionAsync(string sectionName, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Retrieves a secret. If the secret does not exist then an exception is thrown.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the secret.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task<KeyVaultSecret> GetAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -46,31 +58,36 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the secret.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task<KeyVaultSecret> GetIfExistsAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Retrieves a secret's value. If the secret does not exist then an exception is thrown.
+        /// Retrieves a secret's current value. If the secret does not exist then an exception is thrown.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the secret's current value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task<string> GetValueAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Retrieves a secret's value. If the secret does not exist then null is returned.
+        /// Retrieves a secret's current value. If the secret does not exist then null result is returned.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the secret's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task<string> GetValueIfExistsAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Retrieves a collection of secret's names and values. If any secret does not exist then it's value will be null.
+        /// Retrieves a dictionary of secret's names and current values.
+        /// If any secret does not exist then it's value will be null.
         /// </summary>
         /// <param name="names">Collection of secret's names.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be a dictionary of name values.</returns>
         Task<IDictionary<string, string>> GetValuesIfExistsAsync(IEnumerable<string> names, CancellationToken cancellationToken = default);
 
         #endregion
@@ -78,11 +95,12 @@ namespace ByteDev.Azure.KeyVault.Secrets
         #region Get Deleted
 
         /// <summary>
-        /// Checks whether a secret is soft deleted but yet to be purged.
+        /// Determines whether a secret is soft deleted but yet to be purged.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result true if the secret is deleted; otherwise false.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task<bool> IsDeletedAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -90,7 +108,9 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the deleted secret.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task<DeletedSecret> GetDeletedAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -98,7 +118,8 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <returns>The task object representing the asynchronous operation. Result will be the deleted secret or null.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task<DeletedSecret> GetDeletedIfExistsAsync(string name, CancellationToken cancellationToken = default);
 
         #endregion
@@ -106,14 +127,28 @@ namespace ByteDev.Azure.KeyVault.Secrets
         #region Set
 
         /// <summary>
-        /// Sets a secret's value. If the secret does not exist then it is created.
-        /// If the secret exists then its value is updated.
+        /// Sets a secret's current value. If the secret does not exist then it is created.
+        /// If the secret exists then a new current version is created.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="value">New value for the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task SetValueAsync(string name, string value, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Sets a secret's current value in a idempotent manner.
+        /// If the secret does not exist then it is created.
+        /// If the secret exists and it's current value is different to the new one then a new current version is created and result true will be returned.
+        /// If the secret exists and it's current value is equal to the new value then no set will be performed and result false will be returned.
+        /// </summary>
+        /// <param name="name">Name of the secret.</param>
+        /// <param name="value">New value for the secret.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The task object representing the asynchronous operation. Result will be true if a secret was set; otherwise false.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        Task<bool> SafeSetValueAsync(string name, string value, CancellationToken cancellationToken = default);
 
         #endregion
 
@@ -134,6 +169,8 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// <param name="waitToComplete">Indicates if we should wait for the secret delete operation to complete before returning.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task DeleteAsync(string name, bool waitToComplete, CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -143,14 +180,17 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// <param name="waitToComplete">Indicates if we should wait for the secret delete operation to complete before returning.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
         Task DeleteIfExistsAsync(string name, bool waitToComplete, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Soft deletes a secret and purges it. If the secret does not exist then an exception is thrown.
+        /// Deletes a secret and purges it. If the secret does not exist then an exception is thrown.
         /// </summary>
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task DeleteAndPurgeAsync(string name, CancellationToken cancellationToken = default);
         
         #endregion
@@ -163,6 +203,8 @@ namespace ByteDev.Azure.KeyVault.Secrets
         /// <param name="name">Name of the secret.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> cannot be null or empty.</exception>
+        /// <exception cref="T:ByteDev.Azure.KeyVault.Secrets.SecretNotFoundException">Secret could not be found.</exception>
         Task PurgeAsync(string name, CancellationToken cancellationToken = default);
 
         /// <summary>