+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 1.1.0 - 31 March 2021
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `KeyVaultSecretClient.SafeSetValueAsync`.
+- Added `KeyVaultSecretClient.GetSectionAsync`.
+
+Bug fixes / internal changes:
+- XML doc comment improvements.
+
 ## 1.0.0 - 07 September 2020
 
 Initial version.